lifting-lookup-refresh/requirements.txt
selenium
awswrangler 
boto3
python-dotenv
pandas